+++ CotEditor/Sources/FormatPaneController.swift
@@ -287,6 +287,7 @@ final class FormatPaneController: NSViewController, NSMenuItemValidation, NSTabl
     @IBAction func openSyntaxMappingConflictSheet(_ sender: Any?) {
         
         let viewController = NSViewController.instantiate(storyboard: "SyntaxMappingConflictsView")
+        
         self.presentAsSheet(viewController)
     }
     
@@ -295,8 +296,8 @@ final class FormatPaneController: NSViewController, NSMenuItemValidation, NSTabl
     @IBAction func editSyntaxStyle(_ sender: Any?) {
         
         let styleName = self.targetStyleName(for: sender)
-        
-        guard let viewController = SyntaxEditViewController(style: styleName, mode: .edit) else { return }
+        let viewController = SyntaxEditViewController.instantiate(storyboard: "SyntaxEditView")
+        viewController.setup(style: styleName, mode: .edit)
         
         self.presentAsSheet(viewController)
     }
@@ -306,8 +307,8 @@ final class FormatPaneController: NSViewController, NSMenuItemValidation, NSTabl
     @IBAction func duplicateSyntaxStyle(_ sender: Any?) {
         
         let styleName = self.targetStyleName(for: sender)
-        
-        guard let viewController = SyntaxEditViewController(style: styleName, mode: .copy) else { return }
+        let viewController = SyntaxEditViewController.instantiate(storyboard: "SyntaxEditView")
+        viewController.setup(style: styleName, mode: .copy)
         
         self.presentAsSheet(viewController)
     }
@@ -317,8 +318,8 @@ final class FormatPaneController: NSViewController, NSMenuItemValidation, NSTabl
     @IBAction func createSyntaxStyle(_ sender: Any?) {
         
         let styleName = self.targetStyleName(for: sender)
-        
-        guard let viewController = SyntaxEditViewController(style: styleName, mode: .new) else { return }
+        let viewController = SyntaxEditViewController.instantiate(storyboard: "SyntaxEditView")
+        viewController.setup(style: styleName, mode: .new)
         
         self.presentAsSheet(viewController)
     }