+++ CotEditor/Sources/SyntaxEditViewController.swift
@@ -26,88 +26,40 @@
 
 import Cocoa
 
-
-enum SyntaxEditSheetMode: Int {
-    
-    case edit
-    case copy
-    case new
-}
-
-
-private enum PaneIndex: Int {
-    
-    case keywords
-    case commands
-    case types
-    case attributes
-    case variables
-    case values
-    case numbers
-    case strings
-    case characters
-    case comments
-    case separator1
-    case outline
-    case completion
-    case fileMapping
-    case separator2
-    case styleInfo
-    case validation
-    
-    
-    var title: String {
-        switch self {
-        case .keywords: return "Keywords"
-        case .commands: return "Commands"
-        case .types: return "Types"
-        case .attributes: return "Attributes"
-        case .variables: return "Variables"
-        case .values: return "Values"
-        case .numbers: return "Numbers"
-        case .strings: return "Strings"
-        case .characters: return "Characters"
-        case .comments: return "Comments"
-        case .separator1: return String.separator
-        case .outline: return "Outline Menu"
-        case .completion: return "Completion List"
-        case .fileMapping: return "File Mapping"
-        case .separator2: return String.separator
-        case .styleInfo: return "Style Info"
-        case .validation: return "Syntax Validation"
-        }
-    }
-}
-
-
-
-// MARK: -
-
 final class SyntaxEditViewController: NSViewController, NSTextFieldDelegate, NSTableViewDelegate {
+    
+    enum Mode: Int {
+        
+        case edit
+        case copy
+        case new
+    }
+    
+    
     
     // MARK: Private Properties
     
-    private let mode: SyntaxEditSheetMode
-    private let originalStyleName: String
-    @objc private dynamic var style: NSMutableDictionary
+    @objc private dynamic var menuTitles: [String] = []  // for binding
+    
+    private var mode: Mode = .edit
+    private var originalStyleName: String = ""
+    private let style = NSMutableDictionary()
     @objc private dynamic var message: String?
     @objc private dynamic var isStyleNameValid = true
-    private let isBundledStyle: Bool
-    private let isCustomized: Bool
+    @objc private dynamic var isRestoreble = false
+    private var isBundledStyle = false
     
-    private var viewControllers = [NSViewController?]()
+    private var tabViewController: NSTabViewController?
     
-    @IBOutlet private weak var box: NSBox?
     @IBOutlet private weak var menuTableView: NSTableView?
     @IBOutlet private weak var styleNameField: NSTextField?
-    @IBOutlet private weak var restoreButton: NSButton?
     
     
     
     // MARK: -
     // MARK: Lifecycle
     
-    required init?(style styleName: String, mode: SyntaxEditSheetMode) {
+    func setup(style styleName: String, mode: Mode) {
         
         let manager = SyntaxManager.shared
         let name: String
@@ -126,81 +78,43 @@ final class SyntaxEditViewController: NSViewController, NSTextFieldDelegate, NST
             style = manager.blankSettingDictionary
         }
         self.mode = mode
-        self.style = NSMutableDictionary(dictionary: style)
+        self.style.setDictionary(style)
         self.originalStyleName = name
         
         self.isBundledStyle = manager.isBundledSetting(name: name)
-        self.isCustomized = manager.isCustomizedBundledSetting(name: name)
+        self.isRestoreble = manager.isCustomizedBundledSetting(name: name)
         
         if self.isBundledStyle {
             self.message = "Bundled styles can’t be renamed.".localized
         }
-        
-        super.init(nibName: nil, bundle: nil)
     }
     
     
-    required init?(coder: NSCoder) {
+    /// prepare embeded TabViewController
+    override func prepare(for segue: NSStoryboardSegue, sender: Any?) {
         
-        fatalError("init(coder:) has not been implemented")
+        if let destinationController = segue.destinationController as? NSTabViewController {
+            self.tabViewController = destinationController
+            self.menuTitles = destinationController.tabViewItems.map { $0.label.localized }
+            destinationController.children.forEach { $0.representedObject = self.style }
+        }
     }
     
-    
-    override var nibName: NSNib.Name? {
-        
-        return NSNib.Name("SyntaxEditView")
-    }
-    
-    
-    
-    // MARK: View Controller Methods
     
     /// setup UI
     override func viewDidLoad() {
         
         super.viewDidLoad()
         
-        // setup style name field and restore button
+        // setup style name field
         self.styleNameField?.stringValue = self.originalStyleName
-        
         if self.isBundledStyle {
             self.styleNameField?.drawsBackground = false
             self.styleNameField?.isBezeled = false
             self.styleNameField?.isSelectable = false
             self.styleNameField?.isEditable = false
             self.styleNameField?.isBordered = true
-            self.restoreButton?.isEnabled = self.isCustomized
-        } else {
-            self.restoreButton?.isEnabled = false
         }
-        
-        // setup views
-        var viewControllers = [NSViewController?]()
-        for type in SyntaxType.allCases {
-            if type == .comments { break }
-            viewControllers.append(SyntaxTermsEditViewController(syntaxType: type))
-        }
-        viewControllers.append(NSViewController(nibName: "SyntaxCommentsEditView", bundle: nil))
-        viewControllers.append(nil)  // separator
-        viewControllers.append(NSViewController(nibName: "SyntaxOutlineEditView", bundle: nil))
-        viewControllers.append(NSViewController(nibName: "SyntaxCompletionsEditView", bundle: nil))
-        viewControllers.append(NSViewController(nibName: "SyntaxFileMappingEditView", bundle: nil))
-        viewControllers.append(nil)  // separator
-        viewControllers.append(NSViewController(nibName: "SyntaxInfoEditView", bundle: nil))
-        viewControllers.append(SyntaxValidationViewController())
-        
-        for viewController in viewControllers {
-            guard let viewController = viewController else { continue }  // skip separator
-            
-            viewController.representedObject = self.style
-        }
-        
-        self.viewControllers = viewControllers
-        self.swapView(index: 0)
-        
-        // set accessibility
-        self.box?.setAccessibilityElement(true)
-        self.box?.setAccessibilityRole(.group)
     }
     
     
@@ -228,10 +142,9 @@ final class SyntaxEditViewController: NSViewController, NSTextFieldDelegate, NST
         
         guard let tableView = notification.object as? NSTableView else { return assertionFailure() }
         
-        let row = tableView.selectedRow
-        
         // switch view
-        self.swapView(index: row)
+        self.endEditing()
+        self.tabViewController?.selectedTabViewItemIndex = tableView.selectedRow
     }
     
     
@@ -239,7 +152,7 @@ final class SyntaxEditViewController: NSViewController, NSTextFieldDelegate, NST
     func tableView(_ tableView: NSTableView, shouldSelectRow row: Int) -> Bool {
         
         // separator cannot be selected
-        return (PaneIndex(rawValue: row)!.title != String.separator)
+        return (self.menuTitles[row] != .separator)
     }
     
     
@@ -251,14 +164,8 @@ final class SyntaxEditViewController: NSViewController, NSTextFieldDelegate, NST
         
         guard let style = SyntaxManager.shared.bundledSettingDictionary(name: self.originalStyleName) else { return }
         
-        // discard current editing
-        self.discardEditing()
-        
-        // set new content
         self.style.setDictionary(style)
-        
-        // disable "Restore Defaults" button
-        self.restoreButton?.isEnabled = false
+        self.isRestoreble = false
     }
     
     
@@ -295,11 +202,11 @@ final class SyntaxEditViewController: NSViewController, NSTextFieldDelegate, NST
             return
         }
         
-        // validate syntax and display errors (or proceed if user has already seen the error)
-        let validationController = self.viewControllers[PaneIndex.validation.rawValue] as! SyntaxValidationViewController
-        guard validationController.didValidate || validationController.validateSyntax() else {
+        // validate syntax and display errors
+        guard SyntaxStyleValidator.validate(self.style as! SyntaxManager.StyleDictionary).isEmpty else {
             // show "Validation" pane
-            self.menuTableView?.selectRowIndexes(IndexSet(integer: PaneIndex.validation.rawValue), byExtendingSelection: false)
+            let index = self.tabViewController!.tabViewItems.firstIndex { ($0.identifier as? String) == "validation" }!
+            self.menuTableView?.selectRowIndexes(IndexSet(integer: index), byExtendingSelection: false)
             NSSound.beep()
             return
         }
@@ -323,38 +230,8 @@ final class SyntaxEditViewController: NSViewController, NSTextFieldDelegate, NST
     }
     
     
-    /// discard edit and close editor
-    @IBAction func cancel(_ sender: Any?) {
-        
-        self.discardEditing()
-        self.dismiss(sender)
-    }
-    
-    
     
     // MARK: Private Methods
-    
-    /// menu titles for binding
-    @objc var menuTitles: [String] {
-        
-        return (0...PaneIndex.validation.rawValue)
-            .map { PaneIndex(rawValue: $0)!.title.localized }
-    }
-    
-    
-    /// change pane
-    private func swapView(index: Int) {
-        
-        // finish current editing anyway
-        self.endEditing()
-        
-        guard let view = self.viewControllers[index]?.view else { return }
-        
-        // swap views
-        self.box!.contentView = view
-        self.box!.setAccessibilityLabel(self.menuTitles[index])
-    }
-    
     
     /// validate passed-in style name and return if valid
     @discardableResult