+++ CHANGELOG.md
@@ -8,6 +8,7 @@ Change Log
 ### Improvements
 
 - Change not to highlight occurrences of the selection when there are more than 100.
+- Now, syntax style validation is automatically start when the pane switched to “Valitation”.